update readme and add screenshot

+++ src/app/(DashboardLayout)/product_categories/[id]/page.tsx
@@ -25,8 +25,18 @@ export default function DetailPage({
 
   const getAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
-      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/product_categories/${id}`);
+      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/product_categories/${id}`, {
+        method: 'GET',
+        headers: {
+          'Authorization': 'Bearer ' + user.token,
+          'Content-type': 'application/json'
+        }
+      });
       const currentData = await res.json();
       console.log(currentData);
       if (res.status == 200) setData(currentData);
@@ -38,12 +48,17 @@ export default function DetailPage({
 
   const softDeleteAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
       const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/product_categories/${id}`, {
         method: 'DELETE',
         body: JSON.stringify({}),
         headers: {
-            'Content-type': 'application/json'
+          'Authorization': 'Bearer ' + user.token,
+          'Content-type': 'application/json'
         }
       });
       const currentData = await response.json();
@@ -56,6 +71,10 @@ export default function DetailPage({
 
   const hardDeleteAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
       const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/product_categories/${id}`, {
         method: 'DELETE',
@@ -63,7 +82,8 @@ export default function DetailPage({
           "forceDelete": true
         }),
         headers: {
-            'Content-type': 'application/json'
+          'Authorization': 'Bearer ' + user.token,
+          'Content-type': 'application/json'
         }
       });
       const currentData = await response.json();
