+++ src/app/(DashboardLayout)/warehouses/[id]/page.tsx
@@ -32,8 +32,18 @@ export default function DetailPage({
 
   const getAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
-      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/warehouses/${id}`);
+      const res = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/warehouses/${id}`, {
+        method: 'GET',
+        headers: {
+            'Authorization': 'Bearer ' + user.token,
+            'Content-type': 'application/json'
+        }
+      });
       const currentData = await res.json();
       console.log(currentData);
       if (res.status == 200) setData(currentData);
@@ -45,11 +55,16 @@ export default function DetailPage({
 
   const softDeleteAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
       const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/warehouses/${id}`, {
         method: 'DELETE',
         body: JSON.stringify({}),
         headers: {
+            'Authorization': 'Bearer ' + user.token,
             'Content-type': 'application/json'
         }
       });
@@ -63,6 +78,10 @@ export default function DetailPage({
 
   const hardDeleteAPI = async () => {
     try {
+      const localUser = localStorage.getItem("loginUser");
+      if (!localUser) return;
+      const user = JSON.parse(localUser);
+
       const id = (await params).id;
       const response = await fetch(`${process.env.NEXT_PUBLIC_BACKEND_URL}/warehouses/${id}`, {
         method: 'DELETE',
@@ -70,6 +89,7 @@ export default function DetailPage({
           "forceDelete": true
         }),
         headers: {
+            'Authorization': 'Bearer ' + user.token,
             'Content-type': 'application/json'
         }
       });
